Say what the partial import really does, and clean up past a hang-up
The refusal claimed the import "would change nothing". True when the one
placed column is `id`, and false the moment the proposal places
`display_name` and drops a column beside it — then rows DO change, just
not into what the file said. It says it would leave the rest of every
row behind, which is true either way.

And the cleanup took the request's context, so a caller who hung up
between the store and the refusal left the file behind — the moment the
file is most certainly unwanted is the moment a cancelled context stops
the delete. It runs detached with its own bound.

tools_import.go passed the 500-line cap, and the split is where the
concern already divides: bringing a file IN is the half that stores
before it can know the call will succeed, and every refusal past that
point has to decide what becomes of the file.
